feat: simplify UString
