+++ bundle-utils/CMakeLists.txt
@@ -4,9 +4,10 @@ project(bundle-utils)
 set(CMAKE_CXX_STANDARD 17)
 
 add_library(${PROJECT_NAME} STATIC
-        "src/string/QArrayData.h"
-        "src/string/QArrayData.cpp"
-        "src/string/CommonArrayOps.h"
+        "src/string/UStringData.h"
+        "src/string/UStringData.cpp"
+        "src/string/UStringDataPointer.h"
+        "src/string/UStringDataPointer.cpp"
         "src/string/UString.h"
         "src/string/UString.cpp"
         "src/Path.h"
