+++ bundle-utils/src/string/UStringDataPointer.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Duzhong Chen on 2021/3/24.
+//
+
+#include "UStringDataPointer.h"
